Adding changes following mentor's review
