+++ controllers/errorHandlingcontrollers.js
@@ -5,7 +5,7 @@ function handle404NonExistentPaths(req, res, next) {
 function handlePSQL400err(err, req, res, next) {
   const psql400errors = ["22P02", "23502"];
 
-  if (psql400errors.includes(err.code) || err === "Bad request") {
+  if (psql400errors.includes(err.code)) {
     res.status(400).send({ msg: "Bad request" });
   } else {
     next(err);
