+++ models/topicModels.js
@@ -1,0 +1,15 @@
+const db = require("../db/connection");
+
+function fetchTopics() {
+  return db
+    .query(
+      `
+    SELECT * FROM topics
+    `
+    )
+    .then(({ rows }) => {
+      return rows;
+    });
+}
+
+module.exports = {fetchTopics};
